selective tracks api + example, adjust transcriptions to new api
